Handle the SSE stream backpressure and promisify it
Summary:
Normally, the service is able to read Kafka messages faster than it is
able to deliver them to the client, which can cause memory to build up,
especially in cases where there are a lot of messages to be processed
and the client is notoriously slow. Therefore, build the backpressure
into sending the events. Concretely, do not attempt to read the next
message from Kafka before the current event has been flushed down to the
kernel.

While we are at it, also promisify KafkaSSE properly, so that now most
methods return Promises.

Bug: T199813

Reviewers: Ottomata, elukey, Pchelolo, mobrovac

Reviewed By: mobrovac

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D1085