When pausing during consume, don't reject if finished: let next iteration to do so
I don't see why we'd reject any diffefrently than at the top of the loop.
This pause is just to give the V8 loop a little extra time.  Before
continuing the next iteration.  We should resolve after the pause,
start the next iteration, and allow the if this.is_finished check
at the top of the loop end the iteration.