Fix disableSSEFormatting to write a single line at a time, not chars
If disableSSEFormatting is true (perhaps if you want just JSON event data only),
SSEResponse.send was calling _write with a string instead of an array.
This caused the string to be iterated on, which caused each character
to be passed to res.write individually. Call _write directly with the single string line.